backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
networkx==3.4.2
numpy==2.1.3
anthropic==0.40.0
python-dotenv==1.0.1
pydantic==2.10.3
